+++ src/nodes/Hudu/resources/folders/folders.types.ts
@@ -34,7 +34,6 @@ export const folderFilterMapping: FilterMapping<IFolderPostProcessFilters> = {
 
 export interface IFolderPathResponse extends IDataObject {
   path: string;
-  folders: IFolder[];
 }
 
 export type FolderOperation = 'create' | 'get' | 'getAll' | 'update' | 'delete' | 'getPath';